refactored the buymine and minerupgrade scripts so that the minerupgrade script now sits on the same object as buymine script instead of being created in the buymine script. Changed all the costs to floats.

+++ ClickerMinersAndroid/Assets/Scripts/GlobalItems.cs
@@ -18,9 +18,9 @@ public static class GlobalItems {
             "THIS IS A COOL FEATURE",
             "WELCOME TO OUR COOL GAME"
         };
-
-    public static float CalculateNewPrice(int baseCost, float multiplier, int level)
-    {
-        return (float)(baseCost * (float)Math.Pow(multiplier, level - 1.0f));
-    }
+		
+	public static float CalculateNewPrice(float baseCost, float multiplier, int level)
+	{
+		return (float)(baseCost * (float)Math.Pow(multiplier, level - 1.0f));
+	}
 }
